add diet libc to ci
This patch adds dietlibc to the CI. It comes with yet another
implementation for getprogname() and a test for it too. The test showed
the Windows version wasn't working when dynamically linked, so the fix
is included also.

Many meson.build cleanups come with it.

There's a .gitignore now too.