add reallocarray() and some tests
Sadly, none of the arc4random tests test the system RNG itself. Only the
code around it.

I thought it was a good idea so I made sure the tests call other
functions in the library when I could.

This also changes the CI settings to always run all tests even if the
library doesn't provide the functions because the OS does. It's there to
test the tests themselves.