#include_next before all else on headers
I have learned the hard way that placing #include_next after including
other headers is a source of bugs that are hard to debug.

If header A is a libobsd header that extends the system header B, and A
includes an unrelated header C that re-includes A under the belief that
it's including B before A #include_next's B, then C won't have B's
declarations.